Implemented comment and "correspondence" for Workteam
Comments can't be done on task lists, but only on todo-items.

The --ticket switch for teamwork so it's indeterminate. It's a task
list when adding a correspondence, or a task when doing a comment.

Or maybe the logic is off, and the "queue" should be considered the
task list, not the project.