Implemented the RT->TW gateway.
Limitations: no attachment support (yet), but should be doable.
The project must be specified in the configuration.

Please note that the ticket in RT is left untouched and multiple
invocations on the same ticket will append the correspondence to TW
over and over.

Feature: on the fly creation of todo_list via the --queue <name>
argument.