Unified the two scripts into copy-mail-to-rt.pl
The script name is unfortunate and should be changed. EXAMPLES.txt has
been updated. To change source and target it's enough to pass --source
and --target with a string referenced in the configuration file.

Anyway, setting as source Teamwork is not going to work (yet) and same
is true for IMAP as target.